requirements.txt
numpy==1.24.4
pandas==1.5.3
numba==0.57.1
llvmlite==0.40.1
scipy==1.10.1
vectorbt==0.24.3
plotly==5.13.1
streamlit==1.36.0
yfinance==0.2.66
requests
faiss-cpu